refactor: merge lazy loading
