+++ messages/en.json
@@ -15,8 +15,7 @@
     "welcome": "Welcome, ",
     "about": "About the Project",
     "description": "This application was created during the RS School React Course by the Boys-Not-Found team. It demonstrates how to work with RESTful APIs, handle requests, manage application state, and provide useful tools for developers. The project reflects our growth as developers and our passion for learning modern web technologies.",
-    "stack": "Tech Stack: Next.js, Tailwind CSS, Zustand, Firebase.",
-    "loginRequired": "Please log in"
+    "stack": "Tech Stack: Next.js, Tailwind CSS, Zustand, Firebase."
   },
   "navbar": { "history": "History", "variables": "Variables", "rest-client": "Rest-client" },
   "auth": {
@@ -65,6 +64,7 @@
   },
   "history": {
     "title": "History",
-    "no-requests": "No requests yet"
+    "no-requests": "No requests yet",
+    "loginRequired": "Please log in"
   }
 }
