+++ messages/ru.json
@@ -65,6 +65,7 @@
   },
   "history": {
     "title": "История",
-    "no-requests": "Пока нет запросов"
+    "no-requests": "Пока нет запросов",
+    "loginRequired": "Пожалуйста, войдите в систему"
   }
 }
