+++ src/app/[locale]/(privat)/history/page.tsx
@@ -1,13 +1,21 @@
 import { getTranslations, setRequestLocale } from 'next-intl/server';
-import HistoryList from './HistoryList';
-import type { TypeLocale } from '@/types';
+import { TypeLocale } from '@/types';
+import dynamic from 'next/dynamic';
+import Loader from '@/components/Loader/Loader';
 
-type Props = { params: Promise<{ locale: TypeLocale }> };
+type PropsHistoryPage = {
+  params: Promise<{ locale: TypeLocale }>;
+};
 
-export default async function HistoryPage({ params }: Props) {
+const HistoryContent = dynamic(() => import('./_components/HistoryContent'), {
+  ssr: true,
+  loading: () => <Loader />,
+});
+
+export default async function HistoryPage({ params }: PropsHistoryPage) {
   const { locale } = await params;
   setRequestLocale(locale);
-  const t = await getTranslations('home');
+  const t = await getTranslations('history');
 
   const baseUrl =
     process.env.NEXT_PUBLIC_BASE_URL ?? `http://${process.env.VERCEL_URL ?? 'localhost:3000'}`;
@@ -15,17 +23,16 @@ export default async function HistoryPage({ params }: Props) {
   const res = await fetch(`${baseUrl}/api/history`, {
     cache: 'no-store',
   });
-
-  if (!res.ok) {
-    return <p>{t('loginRequired')}</p>;
-  }
+  // if (!res.ok) {
+  //   return <p>{t('loginRequired')}</p>;
+  // }
 
   const requests = await res.json();
 
   return (
     <>
-      <h1>{t('hello')} history Page</h1>
-      <HistoryList requests={requests} />
+      <h2 className="text-2xl font-bold mb-4">{t('title')}</h2>
+      <HistoryContent requests={requests} params={{ locale }} />
     </>
   );
 }
